+++ src/Catalog/Product.php
@@ -8,6 +8,8 @@ declare(strict_types = 1);
 namespace Wizaplace\SDK\Catalog;
 
 use Psr\Http\Message\UriInterface;
+use Wizaplace\SDK\Division\Division;
+use Wizaplace\SDK\Division\DivisionService;
 use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Image\Image;
 use function theodorejb\polycast\to_float;
@@ -104,6 +106,9 @@ final class Product
     /** @var bool */
     private $infiniteStock;
 
+    /** @var array */
+    private $divisions;
+
     /**
      * @internal
      */
@@ -160,6 +165,15 @@ final class Product
             $this->images = array_map(static function (array $imageData) : Image {
                 return new Image($imageData);
             }, $data['images']);
+        }
+
+        if (isset($data['availableOffers'])) {
+            foreach ($data['availableOffers'] as $productId => $availableOffers) {
+                $this->createDivisions($availableOffers, $productId);
+            }
+            foreach ($this->divisions as $key => $divisions) {
+                $this->divisions[$key] = DivisionService::imbricate($divisions);
+            }
         }
     }
 
@@ -408,5 +422,28 @@ final class Product
     public function hasInfiniteStock(): bool
     {
         return $this->infiniteStock;
+    }
+
+    /**
+     * @return array
+     */
+    public function getDivisions(): array
+    {
+        return $this->divisions;
+    }
+
+    /**
+     * @param array $availableOffers
+     * @param int   $productId
+     */
+    private function createDivisions(array $availableOffers, int $productId)
+    {
+        foreach ($availableOffers as $availableOffer) {
+            $this->divisions[$productId][] = new Division($availableOffer);
+
+            if (isset($availableOffer['children']) && !empty($availableOffer['children'])) {
+                $this->createDivisions($availableOffer['children'], $productId);
+            }
+        }
     }
 }